Sub scriber api

+++ src/controller/MessageController.js
@@ -37,10 +37,13 @@ const getAMessage = async (req, res) => {
   const messageId = req.params.id;
   try {
     const Amessage = await messageModel.findById(messageId);
-    res.status(200).json({
-      message: "Single message",
-      data: Amessage,
-    });
+    if (!Amessage)
+      res.status(404).json({ error: `No Message With this id ${messageId}` });
+    else
+      res.status(200).json({
+        message: "Single message",
+        data: Amessage,
+      });
   } catch (error) {
     res.status(500).json({
       error: "Internal Server error",
@@ -50,16 +53,21 @@ const getAMessage = async (req, res) => {
 const updateMessage = async (req, res) => {
   try {
     const messageId = req.params.id;
-    const UpdateStatus = await messageModel.findByIdAndUpdate(messageId, {
-      name: req.body.name,
-      email: req.body.email,
-      message: req.body.message,
-      date: Date.now(),
-    });
-    res.status(200).json({
-      message: `data has been update successfully`,
-      data: UpdateStatus,
-    });
+
+    if (!(await messageModel.findById(messageId)))
+      res.status(404).json({ error: `No Message with this Id ${messageId}` });
+    else {
+      const UpdateStatus = await messageModel.findByIdAndUpdate(messageId, {
+        name: req.body.name,
+        email: req.body.email,
+        message: req.body.message,
+        date: Date.now(),
+      });
+      res.status(201).json({
+        message: `data has been update successfully`,
+        data: UpdateStatus,
+      });
+    }
   } catch (error) {
     res.status(500).json({
       error: "Internal Server error",
@@ -69,10 +77,14 @@ const updateMessage = async (req, res) => {
 const deleteMessage = async (req, res) => {
   try {
     const messageId = req.params.id;
-    const DeleteResults = await messageModel.findByIdAndDelete(messageId);
-    res.status(201).json({
-      message: `Message has been Delete well`,
-    });
+    if (!(await messageModel.findById(messageId)))
+      res.status(404).json({ error: `No Message with this Id ${messageId}` });
+    else {
+      const DeleteResults = await messageModel.findByIdAndDelete(messageId);
+      res.status(201).json({
+        message: `Message has been Delete well`,
+      });
+    }
   } catch (err) {
     res.status(500).json({
       error: "Internal Server error",
